fix: rename base classes to base elements
