+++ src/components/header/header-link/header-link.ts
@@ -1,6 +1,6 @@
 import { customElement } from 'lit/decorators.js';
 
-import { SbbLinkBaseElement } from '../../core/base-classes';
+import { SbbLinkBaseElement } from '../../core/base-elements';
 import { SbbHeaderActionCommonElementMixin } from '../common';
 
 /**
