+++ src/components/link/common/block-link-common.ts
@@ -2,7 +2,7 @@ import type { CSSResultGroup, TemplateResult } from 'lit';
 import { property } from 'lit/decorators.js';
 import { html } from 'lit/static-html.js';
 
-import type { SbbActionBaseElement } from '../../core/base-classes';
+import type { SbbActionBaseElement } from '../../core/base-elements';
 import type { SbbIconPlacement } from '../../core/interfaces';
 import type { AbstractConstructor } from '../../core/mixins';
 import { SbbIconNameMixin } from '../../icon';
