+++ src/components/navigation/navigation-button/navigation-button.ts
@@ -1,6 +1,6 @@
 import { customElement } from 'lit/decorators.js';
 
-import { SbbButtonBaseElement } from '../../core/base-classes';
+import { SbbButtonBaseElement } from '../../core/base-elements';
 import { SbbNavigationActionCommonElementMixin } from '../common';
 
 /**
