+++ src/components/tag/tag/tag.ts
@@ -2,7 +2,7 @@ import type { CSSResultGroup, PropertyValues, TemplateResult } from 'lit';
 import { html } from 'lit';
 import { customElement, property } from 'lit/decorators.js';
 
-import { SbbButtonBaseElement } from '../../core/base-classes';
+import { SbbButtonBaseElement } from '../../core/base-elements';
 import { SbbConnectedAbortController, SbbSlotStateController } from '../../core/controllers';
 import { EventEmitter } from '../../core/eventing';
 import type {
