+++ src/components/teaser-paid/teaser-paid.ts
@@ -1,7 +1,7 @@
 import { html, type CSSResultGroup, type TemplateResult } from 'lit';
 import { customElement } from 'lit/decorators.js';
 
-import { SbbLinkBaseElement } from '../core/base-classes';
+import { SbbLinkBaseElement } from '../core/base-elements';
 
 import style from './teaser-paid.scss?lit&inline';
